Finished main_game.py
I have finished this game. I do NOT recommend you play it. This game was ment so I can learn how to use pygame. This game has a few bugs and features that make it quite annoying to play. And there isn't even much to do in this game itself. Was made just for practice. DO NOT PLAY IF YOU WANT TO KEEP YOUR SANITY!!!